# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API Documentation",
        "description": "This is the Swagger/OpenAPI documentation for my Laravel app",
        "contact": {
            "email": "you@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register new user",
                "description": "Sign up new users",
                "operationId": "RegisterUsers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ali@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password_confirmation",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "login user",
                "description": "Sign in user",
                "operationId": "LoginUsers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ali@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "profile user",
                "description": "see profile user",
                "operationId": "94a5f4c5f5e5755cf43698cf9bc49e9d",
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/editINFO": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit profile user",
                "description": "edit profile user",
                "operationId": "50c366241588acdb6b422cffa7997575",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "username",
                                        "example": "alalisod"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Aliii"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Edited info!!"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "JWT Authorization header using the Bearer scheme."
            }
        }
    },
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}